backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
atlassian-python-api==3.41.4
langchain==0.0.350
openai==1.3.5
anthropic==0.5.0
python-multipart==0.0.6
pgvector==0.2.3
pydantic==2.5.2
pydantic-settings==2.1.0